move remaining ONS jobs
