+++ projects/_01_ingest/ons_pd/tests/test_validate_postcode_directory_raw_data.py
@@ -2,12 +2,16 @@ import unittest
 
 from unittest.mock import Mock, patch
 
-import jobs.validate_postcode_directory_raw_data as job
-
-from tests.test_file_data import ValidatePostcodeDirectoryRawData as Data
-from tests.test_file_schemas import ValidatePostcodeDirectoryRawData as Schemas
-
+import projects._01_ingest.ons_pd.jobs.validate_postcode_directory_raw_data as job
+from projects._01_ingest.unittest_data.ingest_test_file_data import (
+    ValidatePostcodeDirectoryRawData as Data,
+)
+from projects._01_ingest.unittest_data.ingest_test_file_schemas import (
+    ValidatePostcodeDirectoryRawData as Schemas,
+)
 from utils import utils
+
+PATCH_PATH = "projects._01_ingest.ons_pd.jobs.validate_postcode_directory_raw_data"
 
 
 class ValidatePostcodeDirectoryRawDatasetTests(unittest.TestCase):
@@ -30,8 +34,8 @@ class MainTests(ValidatePostcodeDirectoryRawDatasetTests):
     def setUp(self) -> None:
         return super().setUp()
 
-    @patch("utils.utils.write_to_parquet")
-    @patch("utils.utils.read_from_parquet")
+    @patch(f"{PATCH_PATH}.utils.write_to_parquet")
+    @patch(f"{PATCH_PATH}.utils.read_from_parquet")
     def test_main_runs(
         self,
         read_from_parquet_patch: Mock,
