+++ projects/_01_ingest/unittest_data/ingest_test_file_data.py
@@ -497,7 +497,7 @@ class CapacityTrackerNonResData:
 
 
 @dataclass
-class ONSData:
+class IngestONSData:
     sample_rows = [
         ("Yorkshire & Humber", "Leeds", "50.10101"),
         ("Yorkshire & Humber", "York", "52.10101"),
@@ -509,6 +509,50 @@ class ONSData:
         ("Yorkshire & Humber", "York", "52.10101"),
         ("Yorkshire & Humber", "Hull", "53.10101"),
     ]
+
+
+@dataclass
+class ValidatePostcodeDirectoryRawData:
+    raw_postcode_directory_rows = [
+        ("AB1 2CD", "20240101", "cssr", "region", "rui"),
+        ("AB2 2CD", "20240101", "cssr", "region", "rui"),
+        ("AB1 2CD", "20240201", "cssr", "region", "rui"),
+        ("AB2 2CD", "20240201", "cssr", "region", "rui"),
+    ]
+
+
+@dataclass
+class CleanONSData:
+    # fmt: off
+    ons_sample_rows_full = [
+        ("AB10AA", "cssr1", "region1", "subicb1", "icb1", "icb_region1", "ccg1", "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2022", "01", "01", "20220101"),
+        ("AB10AB", "cssr1", "region1", "subicb1", "icb1", "icb_region1", "ccg1", "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2022", "01", "01", "20220101"),
+        ("AB10AA", "cssr2", "region1", "subicb2", "icb2", "icb_region2", None, "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2023", "01", "01", "20230101"),
+        ("AB10AB", "cssr2", "region1", "subicb2", "icb2", "icb_region2", None, "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2023", "01", "01", "20230101"),
+        ("AB10AC", "cssr2", "region1", "subicb2", "icb2", "icb_region2", None, "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2023", "01", "01", "20230101"),
+    ]
+    # fmt: on
+
+
+@dataclass
+class ValidatePostcodeDirectoryCleanedData:
+    raw_postcode_directory_rows = [
+        ("AB1 2CD", "20240101"),
+        ("AB2 2CD", "20240101"),
+        ("AB1 2CD", "20240201"),
+        ("AB2 2CD", "20240201"),
+    ]
+
+    # fmt: off
+    cleaned_postcode_directory_rows = [
+        ("AB1 2CD", date(2024, 1, 1), "cssr", "region", date(2024, 1, 9), "cssr", "region", "rui"),
+        ("AB2 2CD", date(2024, 1, 1), "cssr", "region", date(2024, 1, 9), "cssr", "region", "rui"),
+        ("AB1 2CD", date(2024, 1, 9), "cssr", "region", date(2024, 1, 9), "cssr", "region", "rui"),
+        ("AB2 2CD", date(2024, 1, 9), "cssr", "region", date(2024, 1, 9), "cssr", "region", "rui"),
+    ]
+    # fmt: on
+
+    calculate_expected_size_rows = raw_postcode_directory_rows
 
 
 @dataclass
