+++ tests/test_file_data.py
@@ -254,19 +254,6 @@ class CQCProviderData:
 
 
 @dataclass
-class ONSData:
-    # fmt: off
-    ons_sample_rows_full = [
-        ("AB10AA", "cssr1", "region1", "subicb1", "icb1", "icb_region1", "ccg1", "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2022", "01", "01", "20220101"),
-        ("AB10AB", "cssr1", "region1", "subicb1", "icb1", "icb_region1", "ccg1", "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2022", "01", "01", "20220101"),
-        ("AB10AA", "cssr2", "region1", "subicb2", "icb2", "icb_region2", None, "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2023", "01", "01", "20230101"),
-        ("AB10AB", "cssr2", "region1", "subicb2", "icb2", "icb_region2", None, "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2023", "01", "01", "20230101"),
-        ("AB10AC", "cssr2", "region1", "subicb2", "icb2", "icb_region2", None, "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2023", "01", "01", "20230101"),
-    ]
-    # fmt: on
-
-
-@dataclass
 class CapacityTrackerCareHomeData:
     capacity_tracker_care_home_rows = [
         (
@@ -7423,27 +7410,6 @@ class ValidatePIRCleanedData:
 
 
 @dataclass
-class ValidatePostcodeDirectoryCleanedData:
-    # fmt: off
-    raw_postcode_directory_rows = [
-        ("AB1 2CD", "20240101"),
-        ("AB2 2CD", "20240101"),
-        ("AB1 2CD", "20240201"),
-        ("AB2 2CD", "20240201"),
-    ]
-
-    cleaned_postcode_directory_rows = [
-        ("AB1 2CD", date(2024, 1, 1), "cssr", "region", date(2024, 1, 9), "cssr", "region", "rui"),
-        ("AB2 2CD", date(2024, 1, 1), "cssr", "region", date(2024, 1, 9), "cssr", "region", "rui"),
-        ("AB1 2CD", date(2024, 1, 9), "cssr", "region", date(2024, 1, 9), "cssr", "region", "rui"),
-        ("AB2 2CD", date(2024, 1, 9), "cssr", "region", date(2024, 1, 9), "cssr", "region", "rui"),
-    ]
-    # fmt: on
-
-    calculate_expected_size_rows = raw_postcode_directory_rows
-
-
-@dataclass
 class ValidateCleanedIndCqcData:
     # fmt: off
     merged_ind_cqc_rows = [
@@ -7644,17 +7610,6 @@ class ValidatePIRRawData:
         ("1-000000002", "20240109", 10),
     ]
     # fmt: on
-
-
-@dataclass
-class ValidatePostcodeDirectoryRawData:
-    # fmt: off
-    raw_postcode_directory_rows = [
-        ("AB1 2CD", "20240101", "cssr", "region", "rui"),
-        ("AB2 2CD", "20240101", "cssr", "region", "rui"),
-        ("AB1 2CD", "20240201", "cssr", "region", "rui"),
-        ("AB2 2CD", "20240201", "cssr", "region", "rui"),
-    ]
 
 
 @dataclass
